+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## v0.9.0
 - COD9-003: Add Partido for Locations with B-Prefix zipcodes (#57) [388]
+- COD9-004: Select Crosses in Travel Creator (#59) [140]
 
 ## v0.8.0 (2026-04-17)
 - COD8-001: Add Purple Weight for Travels with Crosses (#39) [23]